imind/nodejs-8 - Separating the routing logic from the server launching logic. module.export exercise.
