+++ servers/application_routing.js
@@ -1,48 +1,13 @@
 const http = require('http');
-const fs = require('fs');
 
-const server = http.createServer( (request, response) => {
-    const url = request.url;
-    const method = request.method;
+const routes = require('../routes.js');
 
-    if (url === '/') {
-        response.write(`<html>
-                            <head>
-                                <title>Routing App: Enter a Message</title>
-                            </head>
-                            <body>
-                                <form action="/message" method="POST">
-                                    <input type="text" name="message" id="messageFieldId"/>
-                                    <button type="submit">Send</button>
-                                </form>
-                            </body>
-                        </html>`);
-        return response.end();
-    } else if (url === '/message' && method === 'POST') {
-
-        const requestBody = [];
-        let parsedBody = '';
-
-         // Ingesting the chunks of request body coming on the stream.
-        request.on('data', (chunk) => {
-            console.log('Request Body so far...' + chunk);
-            requestBody.push(chunk);
-        });
-        
-        // When we finally ingested all the chunks in the stream, we can buffer the requestBody.
-        request.on('end',() => {
-            parsedBody = Buffer.concat(requestBody).toString();
-            const message = parsedBody.split('=')[1];
-
-            // We want to persist the message by saving it to a file.
-            fs.writeFile('message-' + new Date().toISOString() + '.txt', message, (error) => {
-                response.statusCode = 302;
-                response.setHeader("location", "/");
-                return response.end();
-            });
-
-        });
-    }
-});
+/* This is pretty darn confusing, we're not passing the request and response objects
+ * and it is because we're only passing a reference to the function, we're not calling the
+ * function ourselves, the function is going to be called by node itself, and as so
+ * he's going to automatically pass the request and response objects, it's kind of a IOC 
+ * inversion of control where the framework itself is taking care of dependencies.
+ */
+const server = http.createServer(routes);
 
 server.listen(3000);
